+++ extra/form/config.ts
@@ -1,4 +1,5 @@
 import type { Block } from 'payload'
+
 import { defaultLexical } from '@/fields/default-lexical'
 
 export const FormBlock: Block = {