+++ src/components/ui/typography.tsx
@@ -7,7 +7,7 @@ interface Props extends React.HTMLAttributes<HTMLHeadingElement> {}
 const H1: React.FC<Props> = ({ className, ...props }) => {
   return (
     <h1
-      className={cn('scroll-m-20 text-4xl font-extrabold tracking-tight lg:text-5xl', className)}
+      className={cn('scroll-m-20 text-4xl font-bold tracking-tight lg:text-5xl', className)}
       {...props}
     />
   )